Python/etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["Wendrew Oliveira <wendrewdevelop@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "^1.3.22"
psycopg2-binary = "^2.8.6"
pandas = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
